+++ HISTORY.rst
@@ -24,6 +24,7 @@ latest (unreleased)
 * Renaming menu entry Customer Invoices to Customer Invoices / Refunds
 * Invoice delivered quantities configuration
 * Configure Swiss fiscal position
+* Add delivery method manager group and reset delivery method onchange SO line
 * Renaming Pick and Pack types to  Reserve & Pack and Freight labeling
 * Add sale terms and conditions on report